+++ packages/dom-snapshot/src/browser/domNodesToCdt.js
@@ -170,7 +170,11 @@ function domNodesToCdt(docNode, baseUrl, log = noop) {
       addOrUpdateAttribute(node.attributes, 'value', elementNode.value);
     }
 
-    if (elementNode.tagName === 'OPTION' && elementNode.parentElement.value === elementNode.value) {
+    if (
+      elementNode.tagName === 'OPTION' &&
+      elementNode.parentElement.selectedOptions &&
+      Array.from(elementNode.parentElement.selectedOptions).indexOf(elementNode) > -1
+    ) {
       addOrUpdateAttribute(node.attributes, 'selected', '');
     }